[fe] feat: API 분리, user API 기초 정의
